Improve options for what to do if equal images are found

+++ tests/setup_images.py
@@ -1,6 +1,7 @@
 import unittest
 import tempfile
 import random
+import shutil
 from pathlib import Path
 from typing import List, Tuple
 
@@ -61,53 +62,67 @@ class SetupImages(unittest.TestCase):
     noisy_file = Path()
     subdir_file = Path()
 
+    to_create = {'jpeg', 'png', 'subdir', 'half'}
+    tolerate_deleted_files = False
+
     @classmethod
     def setUpClass(cls) -> None:
         """create a number of temporary image files"""
         cls.top_directory = Path(tempfile.mkdtemp())
         cls.image_files = []
-        cls.jpeg_file = create_image(
-            Path(tempfile.mkstemp(dir=cls.top_directory, prefix="jpeg_", suffix=".jpg")[1]),
-            cls.width
-        )
-        cls.image_files.append(cls.jpeg_file)
-        cls.png_file = create_image(
-            Path(tempfile.mkstemp(dir=cls.top_directory, prefix="png_", suffix=".png")[1]),
-            cls.width
-        )
-        cls.image_files.append(cls.png_file)
-        # cls.noisy_file = create_image(
-        #     Path(tempfile.mkstemp(dir=cls.top_directory, prefix="noisy_", suffix=".jpg")[1]),
-        #     cls.width
-        # )
-        # fill_image_with_random_pixels(cls.noisy_file)
-        # cls.image_files.append(cls.noisy_file)
-        cls.sub_directory = Path(tempfile.mkdtemp(dir=cls.top_directory))
-        cls.subdir_file = create_image(
-            Path(tempfile.mkstemp(dir=cls.sub_directory, prefix="subdir_", suffix=".jpg")[1]),
-            cls.width
-        )
-        fill_image_with_random_pixels(cls.subdir_file)
-        cls.image_files.append(cls.subdir_file)
-        cls.half_file = create_image(
-            Path(tempfile.mkstemp(dir=cls.top_directory, prefix="test_half_", suffix=".jpg")[1]),
-            cls.width
-        )
-        image = Image(filename=cls.half_file)
-        image.transform('{}x{}'.format(int(cls.width / 2), int(cls.width * 3 / 8)))
-        save(image, cls.half_file)
-        cls.image_files.append(cls.half_file)
+        if 'jpeg' in cls.to_create:
+            cls.jpeg_file = create_image(
+                Path(tempfile.mkstemp(dir=cls.top_directory, prefix="jpeg_", suffix=".jpg")[1]),
+                cls.width
+            )
+            cls.image_files.append(cls.jpeg_file)
+        if 'png' in cls.to_create:
+            cls.png_file = create_image(
+                Path(tempfile.mkstemp(dir=cls.top_directory, prefix="png_", suffix=".png")[1]),
+                cls.width
+            )
+            cls.image_files.append(cls.png_file)
+        if 'subdir' in cls.to_create:
+            cls.sub_directory = Path(tempfile.mkdtemp(dir=cls.top_directory))
+            cls.subdir_file = create_image(
+                Path(tempfile.mkstemp(dir=cls.sub_directory, prefix="subdir_", suffix=".jpg")[1]),
+                cls.width
+            )
+            fill_image_with_random_pixels(cls.subdir_file)
+            cls.image_files.append(cls.subdir_file)
+        if 'half' in cls.to_create:
+            cls.half_file = create_image(
+                Path(
+                    tempfile.mkstemp(dir=cls.top_directory, prefix="test_half_", suffix=".jpg")[1]
+                ),
+                cls.width
+            )
+            image = Image(filename=cls.half_file)
+            image.transform('{}x{}'.format(int(cls.width / 2), int(cls.width * 3 / 8)))
+            save(image, cls.half_file)
+            cls.image_files.append(cls.half_file)
 
     @classmethod
     def tearDownClass(cls) -> None:
         """"remove all the files created by setUp"""
         for file in cls.image_files:
-            file.unlink()
-        cls.sub_directory.rmdir()
+            file.unlink(missing_ok=cls.tolerate_deleted_files)
+        if 'subdir' in cls.to_create:
+            cls.sub_directory.rmdir()
         cls.top_directory.rmdir()
 
     def get_image_files(self) -> List[Path]:
         return sorted(duplicate.files_in_dirs([self.top_directory]))
+
+    def copy_image_file(self, file: Path) -> Path:
+        copied_file = file.with_suffix(".bak")
+        shutil.copyfile(file, copied_file)
+        self.image_files.append(copied_file)
+        return copied_file
+
+    def delete_image_file(self, file: Path) -> None:
+        file.unlink()
+        self.image_files.remove(file)
 
     @staticmethod
     def is_pair_found(element1: Path, element2: Path, matches: List[Tuple[Path, Path]]) -> bool:
